ar, libc, sh: fix new-archive segfault, %s truncation, and $?
ar: rcmd() ignored getaf()'s return, so a new archive left af = -1
and getdir() read the member header from fd -1, a segfault.  Walk the
existing members only when the archive opened.  (Also folds in the
OBJMAGIC -> obj.h MAGIC rename that was already in the tree.)

libc: doprnt.c held the %s length in a uchar, so any string over 255
bytes was cut - make's 383-byte "ar cr" command echoed at 127 chars
while the archiver still saw the full argument list.

sh: $? now expands to the exit status of the last command; without it
"echo $?" read the ? as a glob and answered "No match.".

+++ src/micronix/lib/libc/doprnt.c
@@ -50,7 +50,7 @@ int *		a;
 {
 	char	c, prec;
 	uchar	fill, left, plus;
-	uchar	i;
+	int	i;
 	uchar	base, width, sign, len;
 	uchar	ftype;
 	extern	short _pnum(), _fnum();
